Remove environment variables from environment plugin.
This commit removes the environment variables from the environment
plugin, as they are already covered by the zprofile in the romcoms.

IMO these shouldn't be set in this plugin, as it is loaded by default
and quietly sets defaults when they where not set by the user.
E.g. a user expects nano because their Distro always uses that by
default but now gets thrown into vim.